change from deposits history to bid history and add it to bid_history router
